# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AnimeCard {
  title: string;
  /** @nullable */
  url?: string | null;
  slug: string;
  /** @nullable */
  image?: string | null;
}

export interface HomeData {
  fresh_drops: AnimeCard[];
  on_air: AnimeCard[];
  new_arrivals: AnimeCard[];
  movies: AnimeCard[];
}

export interface HomeResponse {
  success: boolean;
  data: HomeData;
}

export interface SearchResponse {
  success: boolean;
  query: string;
  results: AnimeCard[];
}

export interface FlatEpisode {
  id: string;
  number: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  season: string;
}

export interface SeriesData {
  title: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  description?: string | null;
  genres?: string[];
  /** @nullable */
  is_movie?: boolean | null;
  movie_players?: string[];
  episodes: FlatEpisode[];
}

export interface SeriesResponse {
  success: boolean;
  data: SeriesData;
}

export interface EpisodeData {
  /** @nullable */
  video_player?: string | null;
  /** @nullable */
  m3u8_link?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  next_episode_id?: string | null;
  /** @nullable */
  prev_episode_id?: string | null;
}

export interface EpisodeResponse {
  success: boolean;
  data: EpisodeData;
}

export type SearchAnimeParams = {
q: string;
};