# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AviStream streaming API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: anime
    description: Anime streaming operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /anime/home:
    get:
      operationId: getAnimeHome
      tags: [anime]
      summary: Get home page anime lists
      responses:
        "200":
          description: Home page content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeResponse"

  /anime/search:
    get:
      operationId: searchAnime
      tags: [anime]
      summary: Search anime by query
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"

  /anime/series/{slug}:
    get:
      operationId: getAnimeSeries
      tags: [anime]
      summary: Get anime series or movie details
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Series details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /anime/episode/{episodeId}:
    get:
      operationId: getAnimeEpisode
      tags: [anime]
      summary: Get episode streaming data. Append --dub to episodeId for dubbed version.
      parameters:
        - in: path
          name: episodeId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Episode streaming data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AnimeCard:
      type: object
      properties:
        title:
          type: string
        url:
          type: ["string", "null"]
        slug:
          type: string
        image:
          type: ["string", "null"]
      required:
        - title
        - slug

    HomeData:
      type: object
      properties:
        fresh_drops:
          type: array
          items:
            $ref: "#/components/schemas/AnimeCard"
        on_air:
          type: array
          items:
            $ref: "#/components/schemas/AnimeCard"
        new_arrivals:
          type: array
          items:
            $ref: "#/components/schemas/AnimeCard"
        movies:
          type: array
          items:
            $ref: "#/components/schemas/AnimeCard"
      required:
        - fresh_drops
        - on_air
        - new_arrivals
        - movies

    HomeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/HomeData"
      required:
        - success
        - data

    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/AnimeCard"
      required:
        - success
        - query
        - results

    FlatEpisode:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        title:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        season:
          type: string
      required:
        - id
        - number
        - season

    SeriesData:
      type: object
      properties:
        title:
          type: string
        slug:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        is_movie:
          type: ["boolean", "null"]
        movie_players:
          type: array
          items:
            type: string
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/FlatEpisode"
      required:
        - title
        - episodes

    SeriesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/SeriesData"
      required:
        - success
        - data

    EpisodeData:
      type: object
      properties:
        video_player:
          type: ["string", "null"]
        m3u8_link:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        next_episode_id:
          type: ["string", "null"]
        prev_episode_id:
          type: ["string", "null"]
      required: []

    EpisodeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/EpisodeData"
      required:
        - success
        - data