# lib/api-zod/src/generated/types/animeCard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnimeCard {
  title: string;
  /** @nullable */
  url?: string | null;
  slug: string;
  /** @nullable */
  image?: string | null;
}