# lib/api-zod/src/generated/types/episodeData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EpisodeData {
  /** @nullable */
  video_player?: string | null;
  /** @nullable */
  m3u8_link?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  next_episode_id?: string | null;
  /** @nullable */
  prev_episode_id?: string | null;
}