# lib/api-zod/src/generated/types/episodeResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EpisodeData } from './episodeData';

export interface EpisodeResponse {
  success: boolean;
  data: EpisodeData;
}