# lib/api-zod/src/generated/types/flatEpisode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FlatEpisode {
  id: string;
  number: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  season: string;
}