# lib/api-zod/src/generated/types/homeData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeCard } from './animeCard';

export interface HomeData {
  fresh_drops: AnimeCard[];
  on_air: AnimeCard[];
  new_arrivals: AnimeCard[];
  movies: AnimeCard[];
}