# lib/api-zod/src/generated/types/homeResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HomeData } from './homeData';

export interface HomeResponse {
  success: boolean;
  data: HomeData;
}