# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnimeCard } from './animeCard';

export interface SearchResponse {
  success: boolean;
  query: string;
  results: AnimeCard[];
}