# lib/api-zod/src/generated/types/seriesData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FlatEpisode } from './flatEpisode';

export interface SeriesData {
  title: string;
  /** @nullable */
  slug?: string | null;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  description?: string | null;
  genres?: string[];
  /** @nullable */
  is_movie?: boolean | null;
  movie_players?: string[];
  episodes: FlatEpisode[];
}