# lib/api-zod/src/generated/types/seriesResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeSalt streaming API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeriesData } from './seriesData';

export interface SeriesResponse {
  success: boolean;
  data: SeriesData;
}